requirements.txt
streamlit
pandas
numpy
scikit-surprise
beautifulsoup4
requests
